Move tests
- Move colorfunc tests over from the refactor branch
- Split component scripts into its own
